scrape_researches/uzum_backup/requirements.txt
# Core
psycopg2-binary>=2.9.9
redis>=5.0.0
aiohttp>=3.9.0
pydantic>=2.5.0

# Web scraping
playwright>=1.40.0

# API
fastapi>=0.109.0
uvicorn>=0.25.0

# Task Queue
celery>=5.3.0

# Data processing
python-dateutil>=2.8.2

# Development
python-dotenv>=1.0.0

